// src/app/auth/forgot-user-id/forgot-user-id-enums.enum.ts
export enum ValidationMessages {
    PleaseEnterSecurityQuestionAnswer = "PleaseEnterSecurityQuestionAnswer",
    MaximumAnswerCharacterLimit = "MaximumAnswerCharacterLimit",
    MinimumAnswerCharacterLimit = "MinimumAnswerCharacterLimit",
    UniqueAnswerForQuestions = "UniqueAnswerForQuestions",
    PleaseEnterEmailAddress = "PleaseEnterEmailAddress",
    PleaseEnterLastName = "PleaseEnterLastName",
    PleaseEnterNewPassword = "PleaseEnterNewPassword",
    PleaseConfirmNewPassword = "PleaseConfirmNewPassword",
    PleaseSelectSecurityQuestion = "PleaseSelectSecurityQuestion",
    ConfirmPasswordDoesNotMatch = "ConfirmPasswordDoesNotMatch",
    PleaseEnterAValidEmailAddress ="PleaseEnterAValidEmailAddress",
    SomeErrorOccured="SomeErrorOccurred",
    MinimumCharactersLimitOfPassword = "MinimumCharacterslimitofPassword",
    MaximumCharactersLimitOfPassword = "MaximumCharacterslimitofPassword"
}

export enum StepLabels {
    Verification = 'Verification',
    SecurityQuestions = 'SecurityQuestion'
}

export enum ToastMessages {
    InvalidEmailOrLastName = "InvalidEmailOrLastName",
    XRMUserVerificationFailed = "XRMUserVerificationFailed",
    AccountPendingActivation = 'AccountPendingActivation',
    UserIdSentSuccessfully = 'UserIdSentSuccessfully',
    SecurityAnswerDoesntmatch = 'SecurityAnswerDoesntmatch'
}