// src/app/auth/forgot-password/forgot-password-enums.enum.ts
export enum ValidationMessages {
    UserInvalid = "UserInvalid",
    PleaseEnterSecurityQuestionAnswer = "PleaseEnterSecurityQuestionAnswer",
    MinimumAnswerCharacterLimit = "MinimumAnswerCharacterLimit",
    MaximumAnswerCharacterLimit = "MaximumAnswerCharacterLimit"
}

export enum RequiredStrings{
    EmptyString = '',

}

export enum NavigationPath {
    Login = '/auth/login'
}